+++ packages/react/Select/src/components/Group/index.ts
@@ -1,0 +1,4 @@
+import { Group } from "./Group";
+
+export { Group } from "./Group";
+export default Group;