+++ packages/react/Select/src/index.ts
@@ -1,0 +1,6 @@
+import { Select } from "./Select";
+import "@nimbus-ds/styles/packages/select/index.css";
+
+export { Select } from "./Select";
+export type { SelectProps } from "./select.types";
+export default Select;