+++ packages/styles/src/index.ts
@@ -21,3 +21,4 @@ export { radio } from "./packages/radio";
 export { toggle } from "./packages/toggle";
 export { link } from "./packages/link";
 export { tooltip } from "./packages/tooltip";
+export { select } from "./packages/select";