+++ packages/styles/src/packages/select/index.ts
@@ -1,0 +1,5 @@
+import { wrapper, icon, appearance } from "./select.style.css";
+
+export const select = {
+  style: { wrapper, icon, appearance },
+};